Changed Price-property type from int to double

+++ Application/Skipass/AddSkipassModel.cs
@@ -2,7 +2,7 @@ namespace Application.Skipass;
 
 public sealed class AddSkipassModel
 {
-    public AddSkipassModel(int balance, Guid tariffId, Guid visitorId, bool status)
+    public AddSkipassModel(double balance, Guid tariffId, Guid visitorId, bool status)
     {
         Balance = balance;
         TariffId = tariffId;
@@ -10,7 +10,7 @@ public sealed class AddSkipassModel
         Status = status;
     }
 
-    public int Balance { get; set; }
+    public double Balance { get; set; }
 
     public Guid TariffId { get; set; }
