+++ Web/Contracts/Tariffication/Requests/CreateTarifficationRequest.cs
@@ -2,14 +2,14 @@ namespace Web.Contracts.Tariffication.Requests;
 
 public sealed class CreateTarifficationRequest
 {
-    public CreateTarifficationRequest(int price, Guid tariffId, Guid locationId)
+    public CreateTarifficationRequest(double price, Guid tariffId, Guid locationId)
     {
         Price = price;
         TariffId = tariffId;
         LocationId = locationId;
     }
     
-    public int Price { get; set; }
+    public double Price { get; set; }
     
     public Guid TariffId { get; set; }
